refactor(dedup): hoist identical agent-brief types to @nimbus-dev/sdk (C1)
8 identical type blocks deduped from cli + gateway into sdk/src/agents/brief-types.ts;
divergent cli types (ImpactFinding/GhostFinding/ConflictCollision) kept local.
